ARM: dts-qcom: Disable selinux audit feature in sm8150 chosen
* We want the kernel code to be as complete as possible and not broken.
* Reference Documentation/admin-guide/kernel-parameters.txt (L389)
* Normal startup, no related audit/avc logs in dmesg and logcat.

form: kdrag0n:
security: selinux: don't require auditing

Audit is bad. It affects performance under all circumstances. It's also
dirty.

From LWN: "Andy submitted a patch to fix this particular problem, but he
didn't stop there. He has come to the conclusion that the audit
subsystem is beyond repair, so his patch marks the whole thing as being
broken, making it generally inaccessible. He cited a number of problems
beyond this security issue: it hurts performance even when it is not
being used, it is not (in his mind) reliable, it has problems with
various architectures, and "its approach to freeing memory is
terrifying." All told, Andy said, we're better off without it"

Signed-off-by: kdrag0n <dragon@khronodragon.com>
Change-Id: I8682dab5e9fad9869564713d06a0f317a6640924